requirements.txt
alembic==1.11.1
anyio==3.7.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dnspython==2.4.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
psycopg2==2.9.6
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
WTForms==3.0.1
